build: change checkstyle to google code format, plus adding spotless
We are regularly fighting with codestyles in pullrequests, which is
annoying to the contributors, as well as the maintainers.

There are tools out there, like spotless, which can automate those
processes. We could define an own codestyle, but in fact there is
already one widely used codestyle from google (google java format).

There are plugins to help us within intellij and vscode to follow
this code style. Hence the overall experience can improve by
switching to this defitinion.

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>